bugfix: add disable mouse state for circles in animtion and add missed types for methods
