+++ src/models/app.model.ts
@@ -1,4 +1,5 @@
 export interface ApppModel {
     init(): string;
+    generateCircles(): void
     render(): Promise<void>;
 }
